requirements.txt
ultralytics
yaml
pandas
torch
numpy
matplotlib
pillow
torch
scikit-image
opencv-python
tqdm
torchvision
ray